orchestrator/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
openai>=1.3.0
python-dotenv>=1.0.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.2
pydantic>=2.5.0
tenacity>=8.2.0
protobuf>=4.24.0
